requirements.txt
# Python 3.13 compatibility fix: After installing requirements, run:
# pip install --force-reinstall aiodns pycares
# This upgrades aiodns to 3.6.1+ and pycares to 4.11.0+ to fix AttributeError with pycares.ares_query_a_result
homeassistant>=2025.11.0
colorlog==6.10.1
pip>=21.3.1
ruff==0.14.8
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
